refactor(src): split daemon/tick.rs by responsibility
